Finished E2E Tests

+++ src/api/1 - Services/series.service.ts
@@ -1,5 +1,5 @@
-import { Series } from '@domain/series.entity';
-import { Injectable } from '@nestjs/common';
+import { Series } from '@domain/entities/series.entity';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 
@@ -19,10 +19,32 @@ export class SeriesService {
   }
 
   async updateSeries(seriesToUpdated: Series): Promise<Series> {
+    const seriesFoundById = await this.seriesRepository.findOne({
+      id: seriesToUpdated.id,
+    });
+
+    if (!seriesFoundById) {
+      throw new HttpException(
+        'A Series with the given Id does not exist',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     return await this.seriesRepository.save(seriesToUpdated);
   }
 
   async deleteSeries(seriesId: string): Promise<void> {
+    const seriesFoundById = await this.seriesRepository.findOne({
+      id: seriesId,
+    });
+
+    if (!seriesFoundById) {
+      throw new HttpException(
+        'A Series with the given Id does not exist',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
     await this.seriesRepository.delete(seriesId);
   }
 }
